[ONESHOT-SPECIFIC] Hopefully get the basic "endings" system working correctly, and message boxes.
Note that everything from this commit on should be specific to this
 OneShot-legacy specific branch.

If I haven't reiterated it enough in the relevant files,
 not only is this a modified version of EasyRPG-Player specific to oneshot-legacy,
 this is in part a derivative work of https://github.com/mathewv/oneshot-legacy ,
 to which the license is:

The MIT License

Copyright (c) 2015 Mathew Velasquez (http://mathew.link/)

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.

Thus, both the GPLv3 and the above license apply to the resulting work where applicable.