[ONESHOT-SPECIFIC] Sleep-status: Can sleep, but can't wake up. Player name substitution should work now.
Current high-priority task is to finish the oneshot_ser interface
 so I can make a generic 'use files in Oneshot dir.' oneshot_ser backend
 and start testing off of that.

Immediately after that I need to make waking up work properly.

Notably, the "custom name" system needs a fix.
(This may apply even in the original.)

I've kept it as-is for now, but I have worked out something that should
 run fine without any changes to save format.

Assuming PlayerName_guess isn't used outside of the initial ProphetBot
 conversation (in order to allow the custom name to take effect there),
 the current system should work fine as long as a custom name isn't kept
 cross-save boundary.

The fix I intend to use is to load the custom name into the username buffer,
 when the game indicates that it wants to start patching over dialogue
 with that name.

Of course, there is still the "small" matter of figuring out how to
 determine the player's name in a cross-platform manner.

I'm sticking with "Pancake" for now.